Add option to configure apt dist-upgrade, autoremove, and clean

+++ files/debian-auto-upgrade.sh
@@ -35,10 +35,20 @@ main() {
 
   export DEBIAN_FRONTEND=noninteractive
 
+  # shellcheck disable=SC1091
+  source /etc/auto-upgrade.conf
+
   wait_for_successful_apt_update
   run apt-get upgrade -y
-  run apt-get autoremove --purge -y
-  run apt-get clean -y
+  if [[ $DEBIAN_APT_DIST_UPGRADE -ne 0 ]]; then
+    run apt-get dist-upgrade -y
+  fi
+  if [[ $DEBIAN_APT_AUTOREMOVE -ne 0 ]]; then
+    run apt-get autoremove --purge -y
+  fi
+  if [[ $DEBIAN_APT_CLEAN -ne 0 ]]; then
+    run apt-get clean -y
+  fi
 }
 
 main "$@"
